started implementation of combining signals
When PivPav synthesizes the operators, it connects every port of the
operator to an IO pin of the FPGA. Some operators need a lot of pins,
which makes the build fail with a message saying that not enough
IOB (I/O pins) are available.

This change enables combining of less important outputs. It adds most
of the support, but it doesn't work yet.